make File|GitSourceClient available only in the test environment (#236)
