--- src/FileSourceClient.php
@@ -1,79 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SmartAssert\SourcesClient;
-
-use Psr\Http\Client\ClientExceptionInterface;
-use SmartAssert\ServiceClient\Authentication\BearerAuthentication;
-use SmartAssert\ServiceClient\Client as ServiceClient;
-use SmartAssert\ServiceClient\Exception\CurlExceptionInterface;
-use SmartAssert\ServiceClient\Exception\HttpResponseExceptionInterface;
-use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
-use SmartAssert\ServiceClient\Exception\InvalidResponseDataException;
-use SmartAssert\ServiceClient\Exception\InvalidResponseTypeException;
-use SmartAssert\ServiceClient\Exception\NonSuccessResponseException;
-use SmartAssert\ServiceClient\Exception\UnauthorizedException;
-use SmartAssert\ServiceClient\Payload\UrlEncodedPayload;
-use SmartAssert\ServiceClient\Request;
-use SmartAssert\SourcesClient\Model\FileSource;
-use SmartAssert\SourcesClient\Model\SourceInterface;
-use SmartAssert\SourcesClient\Request\FileSourceRequest;
-use SmartAssert\SourcesClient\Request\RequestInterface;
-
-class FileSourceClient
-{
-    public function __construct(
-        private readonly ServiceClient $serviceClient,
-        private readonly SourceFactory $sourceFactory,
-        private readonly ExceptionFactory $exceptionFactory,
-        private readonly string $baseUrl,
-    ) {
-    }
-
-    /**
-     * @param non-empty-string $token
-     *
-     * @throws ClientExceptionInterface
-     * @throws CurlExceptionInterface
-     * @throws HttpResponseExceptionInterface
-     * @throws InvalidModelDataException
-     * @throws InvalidResponseDataException
-     * @throws InvalidResponseTypeException
-     * @throws UnauthorizedException
-     */
-    public function create(string $token, string $label): FileSource
-    {
-        return $this->handleRequest(new FileSourceRequest('POST', $label), $token);
-    }
-
-    /**
-     * @throws ClientExceptionInterface
-     * @throws CurlExceptionInterface
-     * @throws HttpResponseExceptionInterface
-     * @throws InvalidModelDataException
-     * @throws InvalidResponseDataException
-     * @throws InvalidResponseTypeException
-     * @throws UnauthorizedException
-     */
-    private function handleRequest(RequestInterface $request, string $token): FileSource
-    {
-        $serviceRequest = (new Request('POST', $this->baseUrl . '/file-source'))
-            ->withPayload(new UrlEncodedPayload($request->getPayload()))
-            ->withAuthentication(new BearerAuthentication($token))
-        ;
-
-        try {
-            $response = $this->serviceClient->sendRequestForJson($serviceRequest);
-        } catch (NonSuccessResponseException $e) {
-            throw $this->exceptionFactory->createFromResponse($e->getResponse());
-        }
-
-        $source = $this->sourceFactory->createFileSource($response->getData());
-        if (null === $source) {
-            throw InvalidModelDataException::fromJsonResponse(SourceInterface::class, $response);
-        }
-
-        return $source;
-    }
-}
