--- tests/Functional/Client/FileSourceClient/CreateTest.php
@@ -1,73 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SmartAssert\SourcesClient\Tests\Functional\Client\FileSourceClient;
-
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
-use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
-use SmartAssert\SourcesClient\Exception\ResponseException;
-use SmartAssert\SourcesClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
-use SmartAssert\SourcesClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
-
-class CreateTest extends AbstractFileSourceClientTestCase
-{
-    use InvalidJsonResponseExceptionDataProviderTrait;
-    use NetworkErrorExceptionDataProviderTrait;
-
-    public function testCreateThrowsInvalidModelDataException(): void
-    {
-        $responsePayload = ['key' => 'value'];
-        $response = new Response(
-            400,
-            ['content-type' => 'application/json'],
-            (string) json_encode($responsePayload)
-        );
-
-        $this->mockHandler->append($response);
-
-        try {
-            ($this->createClientActionCallable())();
-            self::fail(InvalidModelDataException::class . ' not thrown');
-        } catch (InvalidModelDataException $e) {
-            self::assertSame(ResponseException::class, $e->class);
-            self::assertSame($response, $e->getHttpResponse());
-            self::assertSame($responsePayload, $e->payload);
-        }
-    }
-
-    public static function clientActionThrowsExceptionDataProvider(): array
-    {
-        return array_merge(
-            static::networkErrorExceptionDataProvider(),
-            static::invalidJsonResponseExceptionDataProvider(),
-        );
-    }
-
-    protected function createClientActionCallable(): callable
-    {
-        return function () {
-            $this->fileSourceClient->create(self::API_KEY, md5((string) rand()));
-        };
-    }
-
-    protected function getExpectedRequestMethod(): string
-    {
-        return 'POST';
-    }
-
-    protected function getClientActionSuccessResponse(): ResponseInterface
-    {
-        return new Response(
-            200,
-            ['content-type' => 'application/json'],
-            (string) json_encode([
-                'id' => md5((string) rand()),
-                'user_id' => md5((string) rand()),
-                'type' => 'file',
-                'label' => md5((string) rand()),
-            ])
-        );
-    }
-}
