--- tests/Functional/Client/GitSourceClient/CreateTest.php
@@ -1,82 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SmartAssert\SourcesClient\Tests\Functional\Client\GitSourceClient;
-
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Message\ResponseInterface;
-use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
-use SmartAssert\SourcesClient\Exception\ResponseException;
-use SmartAssert\SourcesClient\Tests\Functional\DataProvider\InvalidJsonResponseExceptionDataProviderTrait;
-use SmartAssert\SourcesClient\Tests\Functional\DataProvider\NetworkErrorExceptionDataProviderTrait;
-
-class CreateTest extends AbstractGitSourceClientTestCase
-{
-    use InvalidJsonResponseExceptionDataProviderTrait;
-    use NetworkErrorExceptionDataProviderTrait;
-
-    public function testCreateThrowsInvalidModelDataException(): void
-    {
-        $responsePayload = ['key' => 'value'];
-        $response = new Response(
-            400,
-            ['content-type' => 'application/json'],
-            (string) json_encode($responsePayload)
-        );
-
-        $this->mockHandler->append($response);
-
-        try {
-            ($this->createClientActionCallable())();
-            self::fail(InvalidModelDataException::class . ' not thrown');
-        } catch (InvalidModelDataException $e) {
-            self::assertSame(ResponseException::class, $e->class);
-            self::assertSame($response, $e->getHttpResponse());
-            self::assertSame($responsePayload, $e->payload);
-        }
-    }
-
-    public static function clientActionThrowsExceptionDataProvider(): array
-    {
-        return array_merge(
-            static::networkErrorExceptionDataProvider(),
-            static::invalidJsonResponseExceptionDataProvider(),
-        );
-    }
-
-    protected function createClientActionCallable(): callable
-    {
-        return function () {
-            $this->gitSourceClient->create(
-                self::API_KEY,
-                md5((string) rand()),
-                'https://example.com/' . md5((string) rand()) . '.git',
-                '/' . md5((string) rand()),
-                null
-            );
-        };
-    }
-
-    protected function getExpectedRequestMethod(): string
-    {
-        return 'POST';
-    }
-
-    protected function getClientActionSuccessResponse(): ResponseInterface
-    {
-        return new Response(
-            200,
-            ['content-type' => 'application/json'],
-            (string) json_encode([
-                'type' => 'git',
-                'id' => md5((string) rand()),
-                'user_id' => md5((string) rand()),
-                'label' => md5((string) rand()),
-                'host_url' => 'https://example.com/' . md5((string) rand()) . '.git',
-                'path' => '/' . md5((string) rand()),
-                'has_credentials' => false,
-            ])
-        );
-    }
-}
